pyproject.toml
[project]
name = "milvus-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymilvus[model]>=2.6.6",
    "pytest>=9.0.2",
    "ollama",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "datasets>=4.4.2",
    "peft>=0.18.1",
    "flagembedding>=1.3.5",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
]
